+++ src/Parse_UBLOX.cpp
@@ -258,38 +258,38 @@ uint8_t sempUbloxGetMessageId(const SEMP_PARSE_STATE *parse)
 }
 
 // Get the Payload Length
-uint16_t sempUbloxGetPayloadLength(const SEMP_PARSE_STATE *parse)
+size_t sempUbloxGetPayloadLength(const SEMP_PARSE_STATE *parse)
 {
     SEMP_UBLOX_VALUES *scratchPad = (SEMP_UBLOX_VALUES *)parse->scratchPad;
     return scratchPad->payloadLength;
 }
 
 // Get data
-uint8_t sempUbloxGetU1(const SEMP_PARSE_STATE *parse, uint16_t offset)
+uint8_t sempUbloxGetU1(const SEMP_PARSE_STATE *parse, size_t offset)
 {
     return parse->buffer[offset + SEMP_UBLOX_PAYLOAD_OFFSET];
 }
-uint16_t sempUbloxGetU2(const SEMP_PARSE_STATE *parse, uint16_t offset)
+uint16_t sempUbloxGetU2(const SEMP_PARSE_STATE *parse, size_t offset)
 {
     uint16_t data = parse->buffer[offset + SEMP_UBLOX_PAYLOAD_OFFSET];
     data |= ((uint16_t)parse->buffer[offset + SEMP_UBLOX_PAYLOAD_OFFSET + 1]) << 8;
     return data;
 }
-uint32_t sempUbloxGetU4(const SEMP_PARSE_STATE *parse, uint16_t offset)
+uint32_t sempUbloxGetU4(const SEMP_PARSE_STATE *parse, size_t offset)
 {
     uint32_t data = 0;
     for (uint16_t i = 0; i < sizeof(data); i++)
         data |= ((uint32_t)parse->buffer[offset + SEMP_UBLOX_PAYLOAD_OFFSET + i]) << (8 * i);
     return data;
 }
-uint64_t sempUbloxGetU8(const SEMP_PARSE_STATE *parse, uint16_t offset)
+uint64_t sempUbloxGetU8(const SEMP_PARSE_STATE *parse, size_t offset)
 {
     uint64_t data = 0;
     for (uint16_t i = 0; i < sizeof(data); i++)
         data |= ((uint64_t)parse->buffer[offset + SEMP_UBLOX_PAYLOAD_OFFSET + i]) << (8 * i);
     return data;
 }
-int8_t sempUbloxGetI1(const SEMP_PARSE_STATE *parse, uint16_t offset)
+int8_t sempUbloxGetI1(const SEMP_PARSE_STATE *parse, size_t offset)
 {
     union {
         uint8_t unsignedN;
@@ -298,7 +298,7 @@ int8_t sempUbloxGetI1(const SEMP_PARSE_STATE *parse, uint16_t offset)
     unsignedSignedN.unsignedN = sempUbloxGetU1(parse, offset);
     return unsignedSignedN.signedN;
 }
-int16_t sempUbloxGetI2(const SEMP_PARSE_STATE *parse, uint16_t offset)
+int16_t sempUbloxGetI2(const SEMP_PARSE_STATE *parse, size_t offset)
 {
     union {
         uint16_t unsignedN;
@@ -307,7 +307,7 @@ int16_t sempUbloxGetI2(const SEMP_PARSE_STATE *parse, uint16_t offset)
     unsignedSignedN.unsignedN = sempUbloxGetU2(parse, offset);
     return unsignedSignedN.signedN;
 }
-int32_t sempUbloxGetI4(const SEMP_PARSE_STATE *parse, uint16_t offset)
+int32_t sempUbloxGetI4(const SEMP_PARSE_STATE *parse, size_t offset)
 {
     union {
         uint32_t unsignedN;
@@ -316,7 +316,7 @@ int32_t sempUbloxGetI4(const SEMP_PARSE_STATE *parse, uint16_t offset)
     unsignedSignedN.unsignedN = sempUbloxGetU4(parse, offset);
     return unsignedSignedN.signedN;
 }
-int64_t sempUbloxGetI8(const SEMP_PARSE_STATE *parse, uint16_t offset)
+int64_t sempUbloxGetI8(const SEMP_PARSE_STATE *parse, size_t offset)
 {
     union {
         uint64_t unsignedN;
@@ -325,7 +325,7 @@ int64_t sempUbloxGetI8(const SEMP_PARSE_STATE *parse, uint16_t offset)
     unsignedSignedN.unsignedN = sempUbloxGetU8(parse, offset);
     return unsignedSignedN.signedN;
 }
-float sempUbloxGetR4(const SEMP_PARSE_STATE *parse, uint16_t offset)
+float sempUbloxGetR4(const SEMP_PARSE_STATE *parse, size_t offset)
 {
     union {
         uint32_t unsignedN;
@@ -334,7 +334,7 @@ float sempUbloxGetR4(const SEMP_PARSE_STATE *parse, uint16_t offset)
     unsignedFloat.unsignedN = sempUbloxGetU4(parse, offset);
     return unsignedFloat.flt;
 }
-double sempUbloxGetR8(const SEMP_PARSE_STATE *parse, uint16_t offset)
+double sempUbloxGetR8(const SEMP_PARSE_STATE *parse, size_t offset)
 {
     union {
         uint64_t unsignedN;
@@ -343,7 +343,7 @@ double sempUbloxGetR8(const SEMP_PARSE_STATE *parse, uint16_t offset)
     unsignedFloat.unsignedN = sempUbloxGetU8(parse, offset);
     return unsignedFloat.flt;
 }
-const char *sempUbloxGetString(const SEMP_PARSE_STATE *parse, uint16_t offset)
+const char *sempUbloxGetString(const SEMP_PARSE_STATE *parse, size_t offset)
 {
     return (const char *)(&parse->buffer[offset]);
 }
